Update progress-bar.md
In experimenting with QProgress, it looks like the animate property only affects the stripe animation, if this is correct, please accept this minor documentation change. If not, the goal I am trying to achieve is to clarify that animate does not need to be true for the progress point to tween/animate as is changes. Thanks @nothingismagick !